Git init
After a while, I learned some git to actually implement it aa my version control.
I made a test git repo after reading some git manuals and playing around a bit, I think I have enough knowledge to actually do something useful with git and that is exactly what I am doing. Still it is a considerable small project. So it is great to learn git on the fly without risking to mess up something big...

This is the first commit, and it is not to master, as the version of moneybook I currently have on hand is one which still has bugs, the ones I know are minor. But still. This would be version 2.
I am excited about this :D