Fix for erroneous continuation error "Couldn't parse second argument as an integer" on "moneyboook separate FixedAccount ..."
Many times I intend to separate money for a Fixed Account, passing its name as the second parameter. That leads to the error "Couldn't parse second argument as an integer". However the flow does not return.
This is nothing serious, as the program shortcircuits when trying to create a softreference to the second argument. But it looks ugly and messes up the exit status.
In the future I might make the syntax smarter and more flexible, so this kind of error would not occur in the first place, but for now, fixing this is enough.