Fix for erroneous continuation on more than two arguments for injection.
The problem is that invoking `moneybook inject SomeNumber Something` would proceed normally as of the extra argument was not there.
It's not that big of a deal, but it is ugly and it might lead to confusion and errors as one would expect the program to stop you if you give it inadequate data.