atomic_argument_quantity_control: Explanation
For the sake of implementing Natural Language Invokation it is necessary for better control over the amount of arguments given to the command and to each flow.
As the amount of adequate parameters vary drastically, probably the most sensate thing to do is to implement arguemnt quantity control for every individual flow.

So... What the heck does "atomic argument quantity control"?
→ "atomic" is a facncy word for "indivisible". What I mean is that the control of arguments is built into the flows themselves, rather than "done for someone else".
→ "argument quantity control", quantity is similar to "amount". What I mean is that this control of argument is about how many arguments they are. The word "amount" just sounded off.
So... Putting it all together... This branch is about adding "A control of how many arguments each are given to each flow, which is built directly into the flows themselves".

Btw... The word "amount" instead of "quantity" just sounded off to me. As not only means how many, but also, how much. I just couldn't find a word that suited the built-in factor more than "atomic"; I thought about "modular", "integrated", "embedded", etc, but I felt neither of those hit the bar.