ominpresence: Explaination
Right now, if you invoke moneybook from outside its root dir, a purcharse will fail due that it cannot find the Account files, the same goes for many other operations. This design choice origined from making the program as basic as it could be in order to putting it to use afap. Now that I am uncutting some corners, this is a must have.