Reworked 'File_Is_Account' Account method
I feel the word 'rework' is kind of a big word for what actually happened. Although there are some functional changes(read below), it's overall a change of paradigm. Now the working of the method is the same. The thing is that now, by this change, it communicates a different thing to the dev. That it is not intended to be used directly. Now it serves a much atomic or uncoupled role in the big picture. It is going to be used as an accessory.

Changes:
* Now it checks whethee the given file is a regular file
* Now it checks whether the given file is readable