Added absolute-relative categorization for Account methods
Now that the working dir is not necessarily the same as the root dir of moneybook, whether a function functions with a route to an Account file, or with its name has become key. That's why it is important to tell those two apart. And this commit does exactly that.
Right now each method falls under the categories 'relative' or 'absolute' acordingly.