Reworked/fixed 'Read_Account_Share' Account method
Changes:
* Now the method is a true relative method
Before the method took an argument which was stated as the name of an Account, but it was treated as the path to it. Making it very tricky.
Now the argument is stated as an Account name and is treated as such. Making it work as expected.
* Now the method is categorized as relative
Because it is (truly) relative now, it is categorized as such.
* Now the method checks the Account file is an actual Account file