Reworked 'Read_Account' Account method
This rework was not strictly necessary. Although this rework might protect against some unseen bugs. I did not commit to rework this method because I evaluated the remote chance I would encounter them any time soon... Overall, I just reworked this because I wanted to. I saw the method had some sloppy style and I wanted to calm my urges. You can blame me.

Changes:
* Now it checks the file is an Account file
Before it just checked the file was readable. Reading an arbitrary file, would have, in theory, provoqued a bug. If that bug would have propagated into harmfull behaviour is something I don't know tho. But it is better to prevent its existance in the first place.
* Fixed throrical bug: Broken Account share would read to 0
An Account with a non integer share field, in theory, would have passed as it had a share percentage equal to 0.
* Non functional changes
Some changes to the inner nature of the function that do not change what it does.