logging: explanation
Basically I want moneybook to log purchases. I guess injections and all would be fine as well. But right now only purchases matter to me. Because so many times I am not sure if I have already cashed in some purchase I have made. I won't add support for logging treatment. If I want to check, modify, wipe or create a new logging file, I will do so manually.

So I want moneybook to make a log file. In which the purchases are recorded and it says what the purchase was.