Designing the Log File
A will reserve myself from a header line...                       The fields... The amount of money goes there for sure, the time /
date the same thing, about a description...
It would be kinda annoying to describe when I buy a candy or something. I will make it optional, but make the program warn you when
you don't provide one.

Where it will be stored... In the OS's default place for log files. In the case of termux, it will be `/data/data/com.termux/files/usr/var/log/moneybook.log`.

Finally, an example line of the file would be like:
[Purchase] Sat Aug 31 2024 19:40hs, over the Account `selfinvestment` with a cost of `783`: I bought a girlfriend

Although right now I only care about purchases, I added a type-of-log field because it reminds me of pacman's logs, it looks prettiee and also because maybe in the future I will implement further logging.