Abstracted the Log File checks from the Logging itself
I was thinking... My plan was "Now it will try to log the purchase. And if it can't it will warn the user."
But it would be very annoying for it to alter the funds regardless anyway and push you to manually fix it. But also...

I would like to warn the user if the logging can't be done, before commiting the purchase. For example, if there's no log file, I would like the user to be informed before commiting the purchse instead of it failing after trying. It could fail after waiting for the user to accept if something changed, sure, but it's different that checking before trying.

So I will start out by preparing the later. And for the other problem, I will surely try to log first, and if it can be done, then I will ask the user if they want to continue anyway.