Implemented unmessaged logging into purchase flow
After the purchase is commited by the user, and before altering the funds, it will be tried for the purchase to be logged. If it's not possible it will warn the user and ask if they want to continue regardless.

I will reserve the message festure for another commit.
I am unsure about implementing a pre-check of the log file before commiting anything...