Moved the empty message warning to the bottom of the purchase screen
One gets used to ignore the purchase screen, moving the warning down to the final line makes it way easier to see. And you don't need to manually amend the log line out of you not seeing the warning.
That might have been fixed by adding color to the warning message. But that is a completely different feature and also it's not a good excuse to allow a poor placement of the warning message by putting the responsabiliy on another feature.