Designing the injection log entry
What do I want to ve in there?
* The moment of the injection
* The value

What about a message? I mean... It could be useful to track my incomes and to know how much I was paid in a job, or from where that money came from. Yeah, I will do that.
How about the funds of each account? It could be useful to know that in order to roll something back. But I am not sure if I want to do that... Up until now I have never needed to rollback something. But at the same time, the new *and old* funds of each account is an integral piece of information of every injection, that's why it is displayed on screen...
Ok then, just for the sake of registration, I will record the change of the funds of each account in the entry. I will not be lazy.

What format will I use for the funds? I feel using natural language to denote the funds of an account is too wordy; like: the funds of `savings` were `9000` and are now `123321` *money*, the funds of...
Nah. How about this...
`savings`( `9000` > `12321` ), `allowance`( `2345` > `32123` )...
This seems more reasonable.
And right now I don't support currencies. But if I did, it would be as simple as adding its name right after the new funds, and if I were to support multiple currency funds, to add each one following the account name.

So an entry might look like:
 [Injection] Sun Sep 4 2026 13:36hs, with a value of `21000` *money*; `savings`( `9000` > `12321` ), `allowance`( `2345` > `32123` ), `selfinvestment`( `10000` > `20000` ): Burger King payday

Or without a message:
 [Injection] Sun Sep 4 2026 13:36hs, with a value of `21000` *money*;
`savings`( `9000` > `12321` ), `allowance`( `2345` > `32123` ), `selfinvestment`( `10000` > `20000` )