Separated gathering of information from applying changes
My idea is to first try to log the injection and then do it. However, in order to do the logging, I need the information(the names of the Account files and the new and old balances).
However, there's some problems... It goes as: for each account, the Account file(not a symlink), the old and new balances are gathered, the changes are applied, and then that information is discarded just before re-starting the cycle.

So I have have to re-do the whole information process and separate it from the application of changes to the accounts. Doing that on one commit is, beyond evil, not convenient. So I won't.