Preparing the separation for separating the information gathering
One thing I realize now is that I usually do big commits just because of preserving the functionality of the program throughout them; because I want the program to work each commit, while implementing new stuff, and in order to do that, I usually end up build a whole replacement for a part of the program, and implementing it in the same commit.
But if I just don't need to implement it, then I can work on its smaller parts. That's much more like the use I have seen in git repos, but mainly, it is much simple and efficient to work that way. Plus doing stuff like reverting and so is much more efficient.