Tickets system and changes in the logging system
Under the log directory– /data/data/com.termux/files/usr/var/log on teemux– there will be a directory named moneybook. In it, there will be a directory called "Grocery_Tickets". In there, photos of the tickets of groceries will be stored. Following the next formatting:

*  Name format:
   <Int>NumberOfTicket@DateTime
   Example: 24@Fri Jan 31 2025 16:03hs
   The format is generated with the date command, given '+%a %b %e %Y %H:%Mhs' as format.

This means that the place of log file will need to change as well, and is a change in the approach of the whole logging; from very simple in a selfcontained file, to a more structured approach.