Branch explanation: Transaction_Moment
Basically, up until now the moment in which the purchase occurred and the moment in which the purchase was registered were 2 different, undefined things. The only actual data was the moment in each log entry, leaving undefined if the moment of the monetary change and the moment of the record in the program were the same or not.
But now that difference has grown more and more important in orser to track purchases, and to track where ghost money has gone. So I intent to break the aconception of those two things here.