Invokation design; specifying moment of monetary change
The idea is to indicate the date of the monetary change by passing a flag to the program. -d `date` would specify the moment in which the monetary change occurred.
Also, --datetime would be the long equivalent.

Example:
$ moneybook purchase savings 240 "Sex doll" -d "last monday"

So the log datetime would be "now" and the datetime of the monetary change would be the last monday.