Designing the log format
I could use parentheses alongside the record datetime to indicate the monetary change moment. If they're not present, then the moment odf the transaction was not specified, if no hour was provided, then the N/A could replace the hour.
I will NOT design a format in which the fields are justified(every field always fall under the same space / range of characters. Because I want to fet over this asap.

Example:
[Purchase] Mon Jan 27 2025 16:37hs(Sat Jan 25 2025 N/Ahs), over the Account `savings` with a cost of `350`: Food and snaks for mom