Injection_Support: Branch explanation
Basically, when I started working on all of this, I did so without minding about the support for injections from the very started; I knew that I would have to do it, but I didn't consider it when it came to the git history nor branch structure.
For that I could rewrite the history, dividing the branch in both support for purchases and support for injections. However, rewriting history can go wrong in plenty of ways, consume a lot of time and be a sheer pain(mostly when trying to split commits and stuff, but still).
But the key point is that I am in a hurry, and I need to get over this asap. Because of that, I will also just copy the code for purchases and adapt it for injections instead of abstracting it, as that is far more uncertain, requiring far greater designing(both in code, and program architechure) and debugging efforts. So yeah, I will pretty much copy paste tje code for purchases into the inject flow. Also, the fact that in a way, the copy-pasting depends, or build onto the code presented up to this point makes it adequate for this branch to be on top of all the commits created for the purchases.