Re-done argument handling of logging function (not running)
Ok. So for this project I went a bit wild with the style. I basically went "Let's try stuff out and see what I end up liking(as long as it is encapsulated)".
And boy this is awful. I need to change how the non-account-object parameters are handled in order to add a new one.