Fixed bug for Parameter_Is_Account_Object
Surprisingly(for me), the regex was wrong, and it matched the concatenation of all the account objects as well as a single object.
And I have never realized because the account objects were being passed in isooation. Now, unintendedly, the objects are being passed concatenated with § as delimiter for some reason, and that falsely gives ok to the whole string. That used to be the case, now I have fixed the expression, and I understand why it was wrong and why the new one works.