Reworked function invokation and fixed "buggy" optional arguments
The problem was that although I always intended to be able to pass empty strings as the optional arguments, I designed(and documented) the function to NOT pass them when they're not present. Now I have changed that, and I can pass either an empty string, or not pass the arguments anyway. I still have to get them in order tho.