Camera_Wrapper_App: Branch explanation
For some reason I am not being able to just invoke IMAGE_CAPTURE from the cli. Therefore, I will try to create a wrspper thsn can do it or at least give me some insight as for why it doesn't work. This is that, a bare minimum android app to launch image capture.