Initial commit
Basically I have gathered a few basic notions of how to build a bare minimal android app and a few more things off of chatgpt and some android lessons. And thus is what I have got so far.
I haven't really done much, I wanted to find a few stuff out(like remembering java for example) before getting the thing into git. But now it is the time so here we are :)