Added Expense check to the payment logging function
Shellcheck: ✓
Note that I didn't allow shellcheck to check the sourced file simply because I use it from the web and not as a locsl program with access to the project's files.
Tested: ✓

Ok... So during this whole project I have willingly NOT sticked with any particular style. This is the most apparent for coding, and given that bash is so flexible with its grammar and ways of doing things.

But in a lesser degree with the structure and arquitechure of the program as well. And for the functions this has happened as well. Most of them I have tried to do them "standalone". For example, the logging functions all check the validity of the object to be treated(Account or Expense). So they function like standalone commands in that sense. However, they aren't used in a vacuum, and during the flow of the program, the validity of the object is checked beforehand.

And it makes sense because you would think that the logging function should merely do its job and trust the caller. However, because it is possible that the object was valid at the time of the operation but not at the time of logging, it kind of makes sense to double check before logging, and if something goes wrong, handle it gracefully.

The moment when the validity of the object (before you're prompted Yes or No) is decoupled from the moment the operation is to be logged. Between that moment and the log there's an arbitrary amount of time, and the file might have been invalidated during that process. However, I took the moment of prompting as the reference as for how long ago the object was confirmed to be valid. But that is off, because after the operation is accepted, the new state is written into the object, and as a part of that the object is checked to be valid. And the logging occurs right after that check. So one could skip checking that again as, in normal circumstances, the check is redundant because of the minuscule delay between the writing and the logging.

But for the sake of continuity, and just in case I change the order of the sequence or something, I will continue to make the functions "Standalone" and check the files before doing anything.