Corrected argument quantity for payments in the logging function to support transaction moment
Shellcheck: -
Tested: ✓

To support the datetime flag, along with its argument the game changes because the position no longer correlates with the meaning of each parameter.
But for now I will solely correct the amount of parameters to be allowed.
Update: Nevermind. The reason for that is that I intended to use a flag to pass the datetime string. But that's dumb, the reason of the flag is to resemble gnu's invokation style. But this is not the program nor has anything to do with the program's invokation, this is a log function and doesn't need any "niceness" from a user standpoint. So I will pass the argument like any other optional argument.
I will edit the commit that edits the method signature to remove the flag.