Assigned the datetime flag argument (no-comp)
Shellcheck: ✓
Tested: ✓

With the addition of the datetime flag, the assignation of arguments isn't as straight forward.
And I need to make sure the positional arguments' value get bounded to the right variables.
Update: I intended to make a commit assigning all arguments all at once, but it is too complicated and I am getting lost. So I intend to split it up into different commits.