Missing feature: Management of unallocated money, and messages for separations
"As for a message... I don't know... I should check the log and see if I
 have ever added a message to one of these operations.
Update: Yes, I have. "[Separation] Fri Jul 4 2025 13:20hs, `10,000` wer
e separated for the Expense `Rent`( `0` > `10,000` ): From last bi-week
ly salary($150 left from that income)". Although it's not much about the separation itself, but where did the income came from."

Well... The ability of the program to manage the amount of money I don't have allocated(either in accounts or in expenses) is something I would like to have, because it would simplify stuff a whole lot.
Because more or less I would like the log file to be the journal of my finances, I would like to know where the money comes from, and when it came from. And also, it reduces the "debt stack"; like I have to input the money before doing purchases or payments or something, and that sometimes causes me to put everything in standby until I can take charge of everything. So, beyond journaling, it makes it easier for the program to keeo close track of how much money I have available for what purposes.

I mention all of this here because inherently this feature would be tightly linked with separations. As the money you would separate would come from that unalloacted income, and the log would reflect that somehow.
But for now I will NOT implement any of this, and instead, I will continue to have it undefined and uninforced. And add support for messages to separations to make up for the lack of this feature if I need to. Likely if I end up implementing this "Income" feature, the support for messages for separations will go deprecated and never be used again.