Copied the logging function for payments
You see... I intended to build a function for separations. But I realized it is basically the same thing as that for payments but without the transaction moment and replacing "Payment" with "Separation".
So, what if instead of building pretty much the same thing again, I copy the logging function for payments and replace the word "Payment" with "Separation"? Sounds fair, and I want to get over this asap.

So I took the function's body from the commit "Implemented payment message into the logging function", and placed it under the method's signature for logging payments.

In the following commit I will find and replace the word "Payment" with "Separation". And if necessary I will manually replace any imperfection.