Negative_Funds_Purchases: Branch explanation
Basically, I think allowing purchases to be made over Accounts, even when their funds are zero or less is a good idea and will help me out to keep everything up to date, so I will do that.
It will reduce the "debt stack", and make it easier to keep a close eye on my finances.