Context and why to make the change
Basically, I started developing moneybook with the phylosophy of preparing the program to solve *today's* necessities rather than tomorrow's. And if they happen to change, then make the necessary adjustments to make it happen.

When I started developing moneybook I had my first job and was living with dad. I barely purchased anything, so I purchase something(as in a juice box or cookies) every 1 or 2 weeks if at all. Because of that, every purchase, even if it was something insignificant, was very deliberate and throughly thoughtful beforehand.

Because of that- the minscule volume of purchases, and their deliberatedness -, I used the program to evaluate whether I could make the purchase in the firs place, and off of that decide whether I would actually commit it. Nowadays, however, I am in a very different situation. I live on my own, I have a minimal wage so the margin for error is small, and mainly, I have a partner, and I love to pamper him; to buy him delicious foods as well as buy him goodies and stuff for selfinvestment.

That causes me to often spend more moeny than that I actually have for a determined purpose. And that is dangerous, because I have such a small income, spending more than I have for different purposed threatens my financial sustainability.

The thing is that regardless of the program allows me to purchase beyond the funds of an account, that financial debt is real and is no longer prevented by the progfam restriction...

Furthermore... I think it is detrimental for the program to not allow negative funds in purchases. Because the program now serves as a view into the current state of my finances. And I should be plenty of aware of any debt going on. On top of that, it insites me to accumulate stuff I should cash in because the program doesn't allow me to do so until I receive more income, contributing to the "Debt stack".
Which is a lot of tickets pilled up, which I have in stand by.

So, because not being able to input purchases is no longer stopping me from spending the money anyways, at least I should be well aware of how much money I owe and make it easy to be up to date instead of hoarding tickets and bills.