Make_It_Rewarding: Branch Explanation
Basically, I have found myself holding a stack of transactions/operations to input, or been repeatedly missing the chance to use the program.
I have read Atomic Habits from James Clear. He delves into the importance of making an activity rewarding in order to make it more likely to be repeated, but beyond that, to make it desirable makes it so it is not a bother to keeo doing it / do it again...

I would like to make the program more rewarding/desirable to utilize. I would like to make it addictive to use so it would be easy to do the right thing and keep a close eye on my finances.

Now I have a partner and the amount of purchases and stuff I do has skyrocketed, and due to my low income, the threat to my financial sustainability has as well. So I am more compelled and have more of an incentive to use the program more. But still, that kind of motivation depends a whole lot of the moment, your mood, it requires a lot of deliberate thinking and a decent amount of willpower, and also depends that my relationship is going a'right in the current moment. So I am looking for more of a "visceral" style of rewarding. An effortless, reflective sense of reward rather than an effortful, thoughtful reward.

So that's what I am after with this.