Youtube's_Like_Feature: Branch Explanation
I have seen that sometime ago youtube has implemented an interesting feature. Every now and then, when you like a video, the button grows bigger, shows a short, and colorful anikation before returning to its norma
l state.

I have noticed that this was great because before I noticed I found myself liking more videos that I actually enjoyed in general. So it was really cool.

So I was thinking that I could do something like that...
Maybe I could make use of some kind of animation including colors, possibly size or letter style to display a gratifying message when you commit an operation.

I might use a gratifying sound as well, but that would be something else that is NOT inspired by youtube's "new" like animation.

So this branch is to try doing that. Make the end message rewarding to see, inspired in this feature youtube added.