Youtube's like animation has changed
Before it was a cartoonish animation of the hand slightly increasing in size and the thumb slightly streatching back. This while the thing was "camaleon"ing in different stimulating colors.
Now the color thing is more like an explosion of parallel color red piles.

Here is what I think was the original animation: https://youtu.be/_OxhcGIynWk?feature=shared
This one is the one used in shorts, which is simipar to the one used now in videos: https://youtube.com/shorts/XcZKRQYNqwI?feature=shared
Here is the latest one: https://youtube.com/shorts/YVsNEVnVOV0?feature=shared