Checking the tools at hand
Because I am trying to imitate a visual effect, I have to account for the visual means I have...

The program is made to run on a console(TTY), so my means are limited.

I can display text, and update it if necessary; I can show characters with or without color, stuff like bold or italics and some extra stuff, and also I can create animations via replacing characters with other characters(which might have different colors or effects).

Font size and font style(with the exception of italics or bold) are out of the table either because they change the whole screen or are distro dependent and have compñicated set ups.