Enlargement and stretching is out of the table
In most of youtube's like animations the hand grows and stretches.
However, because of the tools I have at hand, changing the font size and font style is out of the table. So both stretching and making the text grow is not possible.

I might be able to get that feel from switching the font to bold to make it appear larger, but little more than that.