Github_Showcasing: Branch explanation
Basically... I have developed this project for myself to be used. Purely for functional purposes. Because of that, there's next to no documentation on the project's github, and the project is fundamentally incomplete from the standpoint of a new user; there's no tutorial, there's no idication or tool about creating the objects the program operates with, it does not ensure the commands it depends on are present or correct...

However, I have come to use it as something to show in order to be hired in IT. And on reddit, it has been brought to my attention that there's more description of what the program is on my cv than in the github itself...

Which made me realize that if I am going to use this as a resource to get hired for, I should make sure it is ready to be both showcased on github, and used by a new user properly.

And this branch aims to do exactly that.