To do list and order of implementation
So... I need to add a good readme, I need to create utilities to manage the objects the program uses, and I need to make sure the program has every command and stuff it depends on before running. Great.

Now, I am reading how to create a good readme, and it requires showing the program in action, a basic set up, etc. Those things require me to have egeryrhing else set up. Then, the sanity check requires the functional aspect is on check; if I make the program check it has everything it needs to run, and then what it needs to run changes I have to go back again...

So... The order would be:
* Add the tools to manage the program objects(and the log)
* Make sure all of the conditions for the program to run are met
* Craft the README

So I will do that.