Github_Showcasing.README: Branch explanation
So... Basically the readme is key because it is likely the only thing people will check out of your project.
So, especially if you're trying to show it off to someone innorder for them to hire you, it should be attractive, selling, informative and projecting a high status.

Rn it is literally nothing. So I should fix it.
However I should get the other things right first in order for the readme tobe the bow to tie everything up.