Github_Showcasing.Object_Management: Branch explanation
Basically, this program operates with objects stored in the hard drive under the program's directory(which is under your homedir), and also with a log file.

The thing is... Up until now, although the program is designed to operate with them, it exposes no support or insight as for creating, deleting, validating such objects. Which is a huge flaw when it comes to be user-friendly.

Because for a new user to use the program, they should read the help message, then delve into the code, figure out what the objects are and what properties and conditions they must meet in order to be accepted by the program.

This was fine up until now simply because I didn't intend this program to be used by anybody but me. And I saw creating insights and tools to do this as a waste of time. But now the context has changed and this is an absolute must.