Structure and placement of the new modes
So... Clearly each type of operation will have its own unvokation mode.
For example, for creating an object, one may:
~/moneybook $ moneybook create account 'My_Account' 50 2000 "My message"

But how do I structure the procedures of those modes?
Well... I could bunch them in the main file, as I did with the purchase, injection, separation and payment modes(somewhere in the future I would really like to rework that btw).

But I don't really like that. I think the modes of the program require their own separate file. Something like ~/moneybook/bin/Operation_Modes.bash

Yeah... Although... It feels a bit off to have a whole file for both modes about object management AND monetary operations such as purchases and injections. Plus, if I were to create a single file for all modes, it would be off for these new ones to be in it, and the old ones to be somewhere else. And I would like to avoid re-allocating those old modes in this branch because itwould be off-topic (literally, topic branch). I still intend to do that sometime though, but not here, and likely I would create a separate file that would be something like: ~/moneybook/bin/Monetary_Operation_Modes.bash

So I guess I will make a file like: ~/moneybook/bin/Object_Management_Modes.bash

Yeah. I will do that.
Update: Wait. I intended to make a dir called "bin" because the files in there would contain pieces or apendixes of the main flow. However... If I aim to import(source) the modes, then they wouldn't be executables but libraries to be used by the main flow...

So... I guess that for the nature of the files to match the name of the directory they're under there's two options:
* Make them actually executables
  I guess this way I could for example make each of the modes into a separate file and instead of sourcing it, executing it either via calling itbor exec-ing it. I have seen this, with arch's pacman I believe.
* Make them libraries
  I could store them under ~/moneybook/lib/Modes/Object_Management_Modes.bash
  But in order for the structure to remain coherent, I should move the methods(Account_Methods, Expense_Methods, and the file for logging) into something like: ~/moneybook/lib/Object_Methods/

I think I will take the later option... Yeah, I will do that.