Add warning for incorrect share among Accounts to creation function
Tested: ✓
Shellcheck: -

Basically, I intended to halt the creation of an Account if the share among all Accounts didn't equal a 100 percent. But after putting a bit of thought into it I realized that is dumb, because when you are creating an Account chances are you're doing it tleither because there's none, or because you aim to redistribute their shares. So, the only time the shares will actually sum a hundred percent is when you're done. So it had no sense.

But I had already done most of the work. So I will leave that block of code at the bottom of the function for now.