Project suspension
You see... I was trying to get done with this peoject rnnin order to leave it in a good state in order to be showcased on github, so I could improve my chances for getting a job in IT (I would not only do that, obviously). I was looking for a new job in both IT and non IT industries. And I was trying to make progress in both. However... I realize that investing in my IT portfolio is a mid to long term thing, likely it will take me around 3 months to get it good to go.
And rn we are in "temporada" here in my country. Which means it is THE time to get another job outside of IT. So, because the IT thing will take a lot of time anyway, I am wasting my chances by doing it rn. And it is hard, because I enjoy doing it, and it feels like progress because at some point it would lead me to a good job, but it is not what I should be doing because it prevents me from doing what will likely get me a job soon or conclude definitely and free me tk do something else.

So I will let this project in standby for a while. And it is extra hard because I haven't even finished the feature I was implementing yet. But it is what must be done.